Fix CRS file granularity
Using the CompuTrainer utility to convert GPS logs to a course profile
results in lots of very small segments where the height differences are
less than 1 metre. These were incorrectly handled as the altitude was
being stored as a long.

In fact the double to long conversion was always rounding down, for
instance 200+0.01 is 200 as expected, but 200-0.01 is 199. Therefore the
course would gradually ratchet downhill with every small data point.

+++ src/ErgFile.cpp
@@ -272,8 +272,8 @@ void ErgFile::parseComputrainer(QString p)
     bool bIsMetric = true;
 
     // running totals for CRS file format
-    long rdist = 0; // running total for distance
-    long ralt = 200; // always start at 200 meters just to prettify the graph
+    double rdist = 0; // running total for distance
+    double ralt = 200; // always start at 200 meters just to prettify the graph
 
     // open the file
     if (p == "" && ergFile.open(QIODevice::ReadOnly | QIODevice::Text) == false) {
